added pipeline restarts. move tests
