+++ main_test.go
@@ -3,58 +3,11 @@ package main
 import (
 	"encoding/json"
 	"fmt"
-	"net/http"
 	"testing"
 	"time"
 
 	"github.com/fullstorydev/hauser/warehouse"
-	"github.com/nishanths/fullstory"
-	"github.com/pkg/errors"
 )
-
-func TestGetRetryInfo(t *testing.T) {
-	testCases := []struct {
-		err           error
-		expDoRetry    bool
-		expRetryAfter time.Duration
-	}{
-		{
-			err:           errors.New("random error!"),
-			expDoRetry:    true,
-			expRetryAfter: defaultRetryAfterDuration,
-		},
-		{
-			err:           fullstory.StatusError{StatusCode: http.StatusTooManyRequests, RetryAfter: 3 * time.Second},
-			expDoRetry:    true,
-			expRetryAfter: 3 * time.Second,
-		},
-		{
-			err:           fullstory.StatusError{StatusCode: http.StatusInternalServerError, RetryAfter: 3 * time.Second},
-			expDoRetry:    true,
-			expRetryAfter: 3 * time.Second,
-		},
-		{
-			err:           fullstory.StatusError{StatusCode: http.StatusServiceUnavailable, RetryAfter: 3 * time.Second},
-			expDoRetry:    true,
-			expRetryAfter: 3 * time.Second,
-		},
-		{
-			err:           fullstory.StatusError{StatusCode: http.StatusNotFound, RetryAfter: 3 * time.Second},
-			expDoRetry:    false,
-			expRetryAfter: defaultRetryAfterDuration,
-		},
-	}
-
-	for i, tc := range testCases {
-		doRetry, retryAfter := getRetryInfo(tc.err)
-		if doRetry != tc.expDoRetry {
-			t.Errorf("expected %t, got %t for doRetry on test case %d", tc.expDoRetry, doRetry, i)
-		}
-		if retryAfter != tc.expRetryAfter {
-			t.Errorf("expected %v, got %v for doRetry on test case %d", tc.expRetryAfter, retryAfter, i)
-		}
-	}
-}
 
 func TestTransformExportJSONRecord(t *testing.T) {
 	testCases := []struct {
